A few changes to the layout of the landing, calculator and output pages to ensure they're mostly with a responsive layout design.
Output page due to time constraints I wasn't able to set the width of the main active svg graph for mobile screens. So the main section of output page is the same width on mobile as it is on desktop. Still usable just not as pretty.

Also went through and reviewed and changed parts of the website that weren't quite meeting accessibility standards, graphs on output are a little unique, but they do have aria-labels now.